+++ Common/Util/Constants.cs
@@ -1,6 +1,6 @@
 namespace Voxel.Common.Util; 
 
 public static class Constants {
-    public const double TicksPerSecond = 20d;
+    public const int TicksPerSecond = 20;
     public const double SecondsPerTick = 1d / TicksPerSecond;
 }
